Do nothing if component instance does not implement `handleClickOutside`
Removing the `handleClickOutside` check on the component's prototype
makes it possible to wrap a class whose instances' `handleClickOutside`
function gets defined during construction time.

This is especially useful when the `handleClickOutside` is a class
instance field (proposed ES7 feature,
https://facebook.github.io/react/blog/2015/01/27/react-v0.13.0-beta-1.html#autobinding)

Now we have to decide whether to let it throw or silently fail if the
component does not implement `handleClickOutside`. We'll just let if
silently fail and leave a note in the README.