Be extremely pendantic with component/instance terminology
renamed `__wrappedComponent` to `__wrappedInstance`, because it
references a component instance. Hopefully no one is using this in the
wild (it has two underscores!) but incase somebody is, this will be
release in a new major version.

Do not use this in the test. Use the new `wrappedRef` prop which can
achieve the same thing. We should not be using "private" variables
anywhere, even in our own tests.